refactor CORS handle SSE to allow mcp-protocol-version header
