+++ plugins/oauth/http_helpers.go
@@ -1,42 +1,18 @@
 package oauth
 
 import (
+	"github.com/centralmind/gateway/cors"
 	"net/http"
 )
 
 // CORSMiddleware applies standard CORS headers to the response
 func CORSMiddleware(handler http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Add CORS headers
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
-		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-
-		// Handle preflight OPTIONS request
-		if r.Method == "OPTIONS" {
-			w.WriteHeader(http.StatusOK)
+		cors.ApplyCORSHeaders(w, "GET, POST")
+		if cors.HandlePreflight(w, r) {
 			return
 		}
-
 		// Call the original handler
 		handler.ServeHTTP(w, r)
 	})
-}
-
-// ApplyCORSHeaders adds the standard CORS headers to a response
-// For handlers that are not wrapped in middleware
-func ApplyCORSHeaders(w http.ResponseWriter, allowedMethods string) {
-	w.Header().Set("Access-Control-Allow-Origin", "*")
-	w.Header().Set("Access-Control-Allow-Methods", allowedMethods+", OPTIONS")
-	w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-}
-
-// HandlePreflight checks if the request is a preflight OPTIONS request and handles it
-// Returns true if the request was handled (caller should return immediately)
-func HandlePreflight(w http.ResponseWriter, r *http.Request) bool {
-	if r.Method == "OPTIONS" {
-		w.WriteHeader(http.StatusOK)
-		return true
-	}
-	return false
 }
